+++ ceammc/ext/src/lfo/lfo_saw.cpp
@@ -24,5 +24,5 @@ void setup_lfo_saw_tilde()
 {
     SoundExternalFactory<LfoSaw> obj("lfo.saw~");
     obj.addMethod("reset", &LfoSaw::m_reset);
-    obj.setXletsInfo({ "freq in hz", "bang: reset phase" }, { "saw in \\[-1,+1\\] range" });
+    obj.setXletsInfo({ "freq in hz", "bang: reset phase" }, { "saw in \\[-1,+1) range" });
 }