+++ ceammc/ext/src/lfo/lfo_tri.cpp
@@ -22,5 +22,5 @@ void setup_lfo_tri_tilde()
 {
     SoundExternalFactory<LfoTri> obj("lfo.tri~");
     obj.addMethod("reset", &LfoTri::m_reset);
-    obj.setXletsInfo({ "freq in hz", "bang: reset phase" }, { "triangle wave in \\[-1,+1\\] range" });
+    obj.setXletsInfo({ "freq in hz", "bang: reset phase" }, { "triangle wave in \\[-1,+1) range" });
 }